% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_longest_path_by_vessel_id.R
\name{getLongestPathByVesselId}
\alias{getLongestPathByVesselId}
\title{The function aimed to find the longest two consecutives observations
for a given vessel}
\usage{
getLongestPathByVesselId(vesselId, maxTimeInterval = NA)
}
\description{
The function aimed to find the longest two consecutives observations
for a given vessel
}
